[CAI-491][CAI-510] Extract chatbot evaluate feature in a new lambda function (#1564)
* extract evaluation from chatbot

* Update evaluation modules

* Update evaluation modules

* Rename file

* Update evaluation modules

* Update lambda function

* Add evaluation test

* Update judge module

* first draft lambda monitor

* Update chatbot module without evaluator

* Update poetry

* Update prompts

* Update queries.py

* Remove evaluator from chatbot

* docker compose

* poetry lock

* dockerfile

* lint

* Update chatbot modules

* Update chatbot evaluation modules

* gitignore

* Updated queries.py

* send message to SQS queue

* send message to SQS queue for evaluate

* move get_final_response to query

* sqs records iteration

* localstack;

* ignore docker volume dir

* remove fastapi background task

* Update chat_generate as async function

* async chat_generate

* changeset

* record body

* rafactory iam policy and new lambda function evaluation

* update env.example

* terraform file with lambda monitor resources

* missed role in assume role step

* renamed lambda monitor into evaluate

* file renamend

* update github action to deploy lambda evaluate

* fix docker file not fount

* missed ecr action in lambda role

* update iam policy deploy chatboat

* fix deploy action

* fix deploy lambda evaluate

* poetry lock

* access key env vars

* json decode sqs event

* retrived_contexts param in Judge evaluate

* Update engine

* add langfuse score error handling

* remove monitor.py

* log langfuse result

* log langfuse result

* add langfuse trace log

* remove langfuse trace log

* readd langfuse trace log

* force float value

* add langfuse log level

* model.py refactoring

* model.py refactoring

* langfuse flush

* fix models

* langfuse score default value

* langfuse value check

* Fix env var for embedder

* Update chatbot and evaluation test

* Update chatbot tests

* Update .changeset/petite-mice-find.md


* remove comments


* remove langfuse log

* remove debug log

---------

+++ apps/chatbot-evaluate/src/modules/models.py
@@ -31,7 +31,7 @@ def get_llm(
     Args:
         provider (str | None): The provider to use for the LLM. Defaults to the value of the `CHB_PROVIDER` environment variable.
         model_id (str | None): The model ID to use for the LLM. Defaults to the value of the `CHB_MODEL_ID` environment variable.
-        temperature (float | None): The temperature to use for the LLM. Defaults to the value of the `CHB_MODEL_TEMPERATURE` environment variable.
+        temperature (int | None): The temperature to use for the LLM. Defaults to the value of the `CHB_MODEL_TEMPERATURE` environment variable.
         max_tokens (int | None): The maximum number of tokens to generate for the LLM. Defaults to the value of the `CHB_MODEL_MAXTOKENS` environment variable.
     Returns:
         LLM: An instance of the LLM class configured with the specified model ID, temperature, and max tokens.
